ref(metrics): move metrics api code locations endpoint to separate file
